Fix names of cache files
Some timezones have a " character, and this doesn't appear in the
created filename (I think due to python magic), so we need to filter
this out.

Also the cache dict for vEvents and vTimezones was referencing UIDs
while it really should reference filenames (as we work on filenames when
finding what we need to output).